refactor of existing RRT planner as a start point to RRT*; added stub for new app
