+++ apps/rrtstar-demo/CMakeLists.txt
@@ -1,0 +1,18 @@
+INCLUDE(../../cmakemodules/AssureCMakeRootFile.cmake) # Avoid user mistake in CMake source directory
+
+PROJECT(rrtstar-demo)
+
+# Define the executable target:
+ADD_EXECUTABLE(${PROJECT_NAME}
+	rrtstar-demo_main.cpp
+    ${MRPT_VERSION_RC_FILE})
+
+# Add the required libraries for linking:
+TARGET_LINK_LIBRARIES(${PROJECT_NAME} ${MRPT_LINKER_LIBS})
+
+# Dependencies on MRPT libraries:
+#  Just mention the top-level dependency, the rest will be detected automatically, 
+#  and all the needed #include<> dirs added (see the script DeclareAppDependencies.cmake for further details)
+DeclareAppDependencies(${PROJECT_NAME} mrpt-gui mrpt-nav)
+DeclareAppForInstall(${PROJECT_NAME})
+
