+++ libs/nav/src/planners/PlannerRRT_SE2_TPS.cpp
@@ -15,14 +15,6 @@
 #include <mrpt/random.h>
 #include <mrpt/system/filesystem.h>
 
-// For 3D log files
-#include <mrpt/opengl/COpenGLScene.h> 
-#include <mrpt/opengl/CGridPlaneXY.h> 
-#include <mrpt/opengl/stock_objects.h>
-#include <mrpt/opengl/CSetOfLines.h>
-#include <mrpt/opengl/CPointCloud.h>
-#include <mrpt/opengl/CText3D.h>
-
 using namespace mrpt::nav;
 using namespace mrpt::utils;
 using namespace mrpt::math;
@@ -30,23 +22,6 @@ using namespace mrpt::poses;
 using namespace std;
 
 MRPT_TODO("Optimize getNearestNode() with KD-tree!")
-
-PlannerRRT_SE2_TPS::TAlgorithmParams::TAlgorithmParams() :
-	robot_shape_circular_radius(0.30),
-	ptg_cache_files_directory("."),
-	goalBias(0.05),
-	maxLength(1.0),
-	minDistanceBetweenNewNodes(0.10),
-	minAngBetweenNewNodes(mrpt::utils::DEG2RAD(15)),
-	ptg_verbose(true),
-	save_3d_log_freq(0)
-{
-	robot_shape.push_back(mrpt::math::TPoint2D(-0.5, -0.5));
-	robot_shape.push_back(mrpt::math::TPoint2D(0.8, -0.4));
-	robot_shape.push_back(mrpt::math::TPoint2D(0.8, 0.4));
-	robot_shape.push_back(mrpt::math::TPoint2D(-0.5, 0.5));
-}
-
 
 PlannerRRT_SE2_TPS::PlannerRRT_SE2_TPS() :
 	m_initialized(false)
@@ -56,82 +31,13 @@ PlannerRRT_SE2_TPS::PlannerRRT_SE2_TPS() :
 /** Load all params from a config file source */
 void PlannerRRT_SE2_TPS::loadConfig(const mrpt::utils::CConfigFileBase &ini, const std::string &sSect)
 {
-	// Robot shape:
-	// ==========================
-	// polygonal shape
-	{
-		// Robot shape is a bit special to load:
-		params.robot_shape.clear(); 
-		const std::string sShape = ini.read_string(sSect,"robot_shape","");
-		if (!sShape.empty())
-		{
-			CMatrixDouble mShape;
-			if (!mShape.fromMatlabStringFormat(sShape))
-				THROW_EXCEPTION_CUSTOM_MSG1("Error parsing robot_shape matrix: '%s'", sShape.c_str());
-			ASSERT_(size(mShape, 1) == 2);
-			ASSERT_(size(mShape, 2) >= 3);
-
-			for (size_t i = 0; i < size(mShape, 2); i++)
-				params.robot_shape.push_back(TPoint2D(mShape(0, i), mShape(1, i)));
-		}
-	}
-	// circular shape
-	params.robot_shape_circular_radius = ini.read_double(sSect, "robot_shape_circular_radius", 0.0);
-
-	// Load PTG tables:
-	// ==========================
-	m_PTGs.clear();
-
-	const size_t PTG_COUNT = ini.read_int(sSect,"PTG_COUNT",0, true );  //load the number of PTGs
-	for ( unsigned int n=0;n<PTG_COUNT;n++ )
-	{
-		// Generate it:
-		const std::string sPTGName = ini.read_string(sSect,format("PTG%u_Type", n ),"", true );
-		m_PTGs.push_back( CParameterizedTrajectoryGeneratorPtr( CParameterizedTrajectoryGenerator::CreatePTG(sPTGName,ini,sSect,format("PTG%u_", n)) ) );
-	}
+	PlannerTPS_VirtualBase::internal_loadConfig_PTG(ini,sSect);
 }
 
 /** Must be called after setting all params (see `loadConfig()`) and before calling `solve()` */
 void PlannerRRT_SE2_TPS::initialize()
 {
-	ASSERTMSG_(!m_PTGs.empty(),"No PTG was defined! At least one must be especified.");
-
-	// Convert to CPolygon for API requisites:
-	mrpt::math::CPolygon poly_robot_shape;
-	poly_robot_shape.clear();
-	if (!params.robot_shape.empty())
-	{
-		vector<double> xm,ym;
-		params.robot_shape.getPlotData(xm,ym);
-		poly_robot_shape.setAllVertices(xm,ym);
-	}
-
-	for (size_t i=0;i<m_PTGs.size();i++)
-	{
-		mrpt::utils::CTimeLoggerEntry tle(m_timelogger, "PTG_initialization");
-
-		// Polygonal robot shape?
-		{
-			mrpt::nav::CPTG_DiffDrive_CollisionGridBased * diff_ptg = dynamic_cast<mrpt::nav::CPTG_DiffDrive_CollisionGridBased *>(m_PTGs[i].pointer());
-			if (diff_ptg) {
-				ASSERTMSG_(!poly_robot_shape.empty(), "No polygonal robot shape specified, and PTG requires one!");
-				diff_ptg->setRobotShape(poly_robot_shape);
-			}
-		}
-		// Circular robot shape?
-		{
-			mrpt::nav::CPTG_RobotShape_Circular *ptg = dynamic_cast<mrpt::nav::CPTG_RobotShape_Circular *>(m_PTGs[i].pointer());
-			if (ptg) {
-				ASSERTMSG_(params.robot_shape_circular_radius>0, "No circular robot shape specified, and PTG requires one!");
-				ptg->setRobotShapeRadius(params.robot_shape_circular_radius);
-			}
-		}
-
-		m_PTGs[i]->initialize(
-			mrpt::format("%s/TPRRT_PTG_%03u.dat.gz", params.ptg_cache_files_directory.c_str(), static_cast<unsigned int>(i)),
-			params.ptg_verbose
-			);
-	}
+	PlannerTPS_VirtualBase::internal_initialize_PTG();
 
 	m_initialized = true;
 }
@@ -478,404 +384,6 @@ void PlannerRRT_SE2_TPS::solve(
 	result.computation_time = working_time.Tac();
 
 }  // end solve()
-
-// Auxiliary function:
-void PlannerRRT_SE2_TPS::transformPointcloudWithSquareClipping(
-	const mrpt::maps::CPointsMap & in_map,
-	mrpt::maps::CPointsMap       & out_map,
-	const mrpt::poses::CPose2D   & asSeenFrom,
-	const double MAX_DIST_XY
-	)
-{
-	size_t nObs;
-	const float *obs_xs, *obs_ys, *obs_zs;
-	in_map.getPointsBuffer(nObs, obs_xs,obs_ys, obs_zs);
-
-	out_map.clear();
-	out_map.reserve(nObs); // Prealloc mem for speed-up
-
-	const CPose2D invPose = -asSeenFrom;
-	// We can safely discard the rest of obstacles, since they cannot be converted into TP-Obstacles anyway!
-
-	for (size_t obs=0;obs<nObs;obs++)
-	{
-		const double gx = obs_xs[obs], gy = obs_ys[obs];
-
-		if (std::abs(gx-asSeenFrom.x())>MAX_DIST_XY || std::abs(gy-asSeenFrom.y())>MAX_DIST_XY)
-			continue;   // ignore this obstacle: anyway, I don't know how to map it to TP-Obs!
-
-		double ox,oy;
-		invPose.composePoint(gx,gy,ox,oy);
-
-		out_map.insertPointFast(ox,oy,0);
-	}
-}
-
-/*---------------------------------------------------------------
-SpaceTransformer
----------------------------------------------------------------*/
-void PlannerRRT_SE2_TPS::spaceTransformer( 
-	const mrpt::maps::CSimplePointsMap &in_obstacles,
-	const mrpt::nav::CParameterizedTrajectoryGenerator *in_PTG,
-	const double MAX_DIST,
-	std::vector<double> &out_TPObstacles
-	)
-{
-	using namespace mrpt::nav;
-	try
-	{
-		// Take "k_rand"s and "distances" such that the collision hits the obstacles
-		// in the "grid" of the given PT
-		// --------------------------------------------------------------------
-		size_t nObs;
-		const float *obs_xs, *obs_ys, *obs_zs;
-		// = in_obstacles.getPointsCount();
-		in_obstacles.getPointsBuffer(nObs, obs_xs, obs_ys, obs_zs);
-
-		// Init obs ranges: 
-		in_PTG->initTPObstacles(out_TPObstacles);
-
-		for (size_t obs=0;obs<nObs;obs++)
-		{
-			const float ox = obs_xs[obs];
-			const float oy = obs_ys[obs];
-
-			if (std::abs(ox)>MAX_DIST || std::abs(oy)>MAX_DIST)
-				continue;   // ignore this obstacle: anyway, I don't know how to map it to TP-Obs!
-
-			in_PTG->updateTPObstacle(ox, oy, out_TPObstacles);
-		}
-
-		// Leave distances in out_TPObstacles un-normalized ([0,1]), so they just represent real distances in meters.
-	}
-	catch (std::exception &e)
-	{
-		cerr << "[PT_RRT::SpaceTransformer] Exception:" << endl;
-		cerr << e.what() << endl;
-	}
-	catch (...)
-	{
-		cerr << "\n[PT_RRT::SpaceTransformer] Unexpected exception!:\n";
-		cerr << format("*in_PTG = %p\n", (void*)in_PTG );
-		if (in_PTG)
-			cerr << format("PTG = %s\n",in_PTG->getDescription().c_str());
-		cerr << endl;
-	}
-}
-
-void PlannerRRT_SE2_TPS::spaceTransformerOneDirectionOnly(
-	const int tp_space_k_direction,
-	const mrpt::maps::CSimplePointsMap &in_obstacles,
-	const mrpt::nav::CParameterizedTrajectoryGenerator *in_PTG,
-	const double MAX_DIST,
-	double &out_TPObstacle_k
-)
-{
-	using namespace mrpt::nav;
-	try
-	{
-		// Take "k_rand"s and "distances" such that the collision hits the obstacles
-		// in the "grid" of the given PT
-		// --------------------------------------------------------------------
-		size_t nObs;
-		const float *obs_xs, *obs_ys, *obs_zs;
-		// = in_obstacles.getPointsCount();
-		in_obstacles.getPointsBuffer(nObs, obs_xs, obs_ys, obs_zs);
-
-		// Init obs ranges: 
-		in_PTG->initTPObstacleSingle(tp_space_k_direction, out_TPObstacle_k);
-
-		for (size_t obs = 0; obs<nObs; obs++)
-		{
-			const float ox = obs_xs[obs];
-			const float oy = obs_ys[obs];
-
-			if (std::abs(ox)>MAX_DIST || std::abs(oy)>MAX_DIST)
-				continue;   // ignore this obstacle: anyway, I don't know how to map it to TP-Obs!
-
-			in_PTG->updateTPObstacleSingle(ox, oy, tp_space_k_direction, out_TPObstacle_k);
-		}
-
-		// Leave distances in out_TPObstacles un-normalized ([0,1]), so they just represent real distances in meters.
-	}
-	catch (std::exception &e)
-	{
-		cerr << "[PT_RRT::SpaceTransformer] Exception:" << endl;
-		cerr << e.what() << endl;
-	}
-	catch (...)
-	{
-		cerr << "\n[PT_RRT::SpaceTransformer] Unexpected exception!:\n";
-		cerr << format("*in_PTG = %p\n", (void*)in_PTG);
-		if (in_PTG)
-			cerr << format("PTG = %s\n", in_PTG->getDescription().c_str());
-		cerr << endl;
-	}
-}
-
-
-void PlannerRRT_SE2_TPS::renderMoveTree(
-	mrpt::opengl::COpenGLScene &scene,
-	const PlannerRRT_SE2_TPS::TPlannerInput  &pi,
-	const PlannerRRT_SE2_TPS::TPlannerResult &result,
-	const TRenderPlannedPathOptions &options
-	)
-{
-	// Build a model of the vehicle shape: 
-	mrpt::opengl::CSetOfLinesPtr gl_veh_shape = mrpt::opengl::CSetOfLines::Create();
-	double xyzcorners_scale; // Size of XYZ corners (scaled to vehicle dimensions)
-	{
-		gl_veh_shape->setLineWidth(options.vehicle_line_width);
-		gl_veh_shape->setColor_u8(options.color_vehicle );
-		
-		double max_veh_radius = 0.;
-		if (!params.robot_shape.empty())
-		{
-			gl_veh_shape->appendLine(
-				params.robot_shape[0].x, params.robot_shape[0].y, 0,
-				params.robot_shape[1].x, params.robot_shape[1].y, 0);
-			for (size_t i = 2; i <= params.robot_shape.size(); i++)
-			{
-				const size_t idx = i % params.robot_shape.size();
-				mrpt::utils::keep_max(max_veh_radius, params.robot_shape[idx].norm());
-				gl_veh_shape->appendLineStrip(params.robot_shape[idx].x, params.robot_shape[idx].y, 0);
-			}
-		} 
-		else if (params.robot_shape_circular_radius>0)
-		{
-			const int NUM_VERTICES = 10;
-			const double R = params.robot_shape_circular_radius;
-			for (int i = 0; i <= NUM_VERTICES; i++)
-			{
-				const size_t idx  = i % NUM_VERTICES;
-				const size_t idxn = (i+1) % NUM_VERTICES;
-				const double ang  = idx  * 2 * M_PI / (NUM_VERTICES - 1);
-				const double angn = idxn * 2 * M_PI / (NUM_VERTICES - 1);
-				gl_veh_shape->appendLine(
-					R*cos(ang), R*sin(ang), 0,
-					R*cos(angn), R*sin(angn), 0);
-			}
-			mrpt::utils::keep_max(max_veh_radius, R);
-		}
-
-		xyzcorners_scale = max_veh_radius * 0.5;
-	}
-	// Override with user scale?
-	if (options.xyzcorners_scale!=0)
-		xyzcorners_scale = options.xyzcorners_scale; 
-
-	// "ground"
-	if (options.ground_xy_grid_frequency>0)
-	{
-		mrpt::opengl::CGridPlaneXYPtr obj = mrpt::opengl::CGridPlaneXY::Create( pi.world_bbox_min.x,pi.world_bbox_max.x, pi.world_bbox_min.y,pi.world_bbox_max.y, 0, options.ground_xy_grid_frequency );
-		obj->setColor_u8(options.color_ground_xy_grid);
-		scene.insert( obj );
-	}
-
-	// Original randomly-pick pose:
-	if (options.x_rand_pose)
-	{
-		mrpt::opengl::CSetOfObjectsPtr obj = mrpt::opengl::stock_objects::CornerXYZ(xyzcorners_scale*1.0);
-		string m_name =	"X_rand";
-		obj->setName(m_name);
-		obj->enableShowName();
-		obj->setPose(*options.x_rand_pose);
-		scene.insert(obj);
-	}
-
-	// Nearest state pose:
-	if (options.x_nearest_pose)
-	{
-		mrpt::opengl::CSetOfObjectsPtr obj = mrpt::opengl::stock_objects::CornerXYZ(xyzcorners_scale*1.0);
-		string m_name =	"X_near";
-		obj->setName(m_name);
-		obj->enableShowName();
-		obj->setPose(*options.x_nearest_pose);
-		scene.insert(obj);
-	}
-
-	// Determine the up-to-now best solution, so we can highlight the best path so far:
-	TMoveTreeSE2_TP::path_t  best_path;
-	if (options.highlight_path_to_node_id!=INVALID_NODEID)
-		result.move_tree.backtrackPath( options.highlight_path_to_node_id, best_path);
-
-	// make list of nodes in the way of the best path:
-	std::set<const TMoveEdgeSE2_TP*> edges_best_path;
-	std::set<const TMoveEdgeSE2_TP*> edges_best_path_decim;
-	if (!best_path.empty())
-	{
-		TMoveTreeSE2_TP::path_t::const_iterator it_end = best_path.end();
-		TMoveTreeSE2_TP::path_t::const_iterator it_end_1 = best_path.end(); std::advance(it_end_1,-1);
-
-		for (TMoveTreeSE2_TP::path_t::const_iterator it=best_path.begin();it!=it_end;++it)
-			if (it->edge_to_parent)
-				edges_best_path.insert(it->edge_to_parent);
-
-		// Decimate the path (always keeping the first and last entry):
-		ASSERT_ABOVE_(options.draw_shape_decimation,0)
-		for (TMoveTreeSE2_TP::path_t::const_iterator it=best_path.begin();it!=it_end; )
-		{
-			if (it->edge_to_parent)
-				edges_best_path_decim.insert(it->edge_to_parent);
-			if (it==it_end_1) break;
-			for (size_t k=0;k<options.draw_shape_decimation;k++) {
-				if (it==it_end || it==it_end_1) break;
-				++it;
-			}
-		}
-	}
-
-	// The starting pose vehicle shape must be inserted independently, because the rest are edges and we draw the END pose of each edge:
-	{
-		mrpt::opengl::CSetOfLinesPtr vehShape(new mrpt::opengl::CSetOfLines(*gl_veh_shape));
-		mrpt::poses::CPose3D shapePose(mrpt::math::TPose3D(pi.start_pose));
-		shapePose.z_incr( options.vehicle_shape_z );
-		vehShape->setPose(shapePose);
-		scene.insert(vehShape);
-	}
-
-	// Existing nodes & edges between them:
-	{
-		const TMoveTreeSE2_TP::node_map_t & lstNodes = result.move_tree.getAllNodes();
-
-		for (TMoveTreeSE2_TP::node_map_t::const_iterator itNode=lstNodes.begin();itNode!=lstNodes.end();++itNode)
-		{
-			const TMoveTreeSE2_TP::NODE_TYPE & node = itNode->second;
-
-			mrpt::poses::CPose2D parent_state; 
-			if (node.parent_id!=INVALID_NODEID) 
-			{
-				parent_state=  lstNodes.find(node.parent_id)->second.state;
-			}
-			const mrpt::poses::CPose2D trg_state = node.state;
-
-			const bool is_new_one = (itNode==(lstNodes.end()-1));
-			const bool is_best_path = edges_best_path.count(node.edge_to_parent)!=0;
-			const bool is_best_path_and_draw_shape = edges_best_path_decim.count(node.edge_to_parent)!=0;
-
-			// Draw children nodes:
-			{
-				const float corner_scale = xyzcorners_scale* (is_new_one ? 1.5f : 1.0f);
-
-				mrpt::opengl::CSetOfObjectsPtr obj = mrpt::opengl::stock_objects::CornerXYZSimple(corner_scale);
-				obj->setPose(trg_state);
-				scene.insert(obj);
-
-				// Insert vehicle shapes along optimal path:
-				if (is_best_path_and_draw_shape) 
-				{
-					mrpt::opengl::CSetOfLinesPtr vehShape(new mrpt::opengl::CSetOfLines(*gl_veh_shape));
-					mrpt::poses::CPose3D shapePose = trg_state;
-					shapePose.z_incr( options.vehicle_shape_z );
-					vehShape->setPose(shapePose);
-					scene.insert(vehShape);
-				}
-			}
-
-			// Draw line parent -> children nodes.
-			if (node.parent_id!=INVALID_NODEID) 
-			{
-				// Draw actual PT path between parent and children nodes:
-				ASSERT_(node.edge_to_parent)
-				const mrpt::nav::CParameterizedTrajectoryGenerator * ptg = m_PTGs[node.edge_to_parent->ptg_index].pointer();
-
-				// Create the path shape, in relative coords to the parent node:
-				mrpt::opengl::CSetOfLinesPtr obj = mrpt::opengl::CSetOfLines::Create();
-				obj->setPose(parent_state); // Points are relative to this pose: let OpenGL to deal with the coords. composition
-
-				ptg->renderPathAsSimpleLine(node.edge_to_parent->ptg_K,*obj,0.25f /*decimation*/, node.edge_to_parent->ptg_dist /*max path length*/);
-
-				if (is_new_one && options.highlight_last_added_edge)
-				{
-					// Last edge format:
-					obj->setColor_u8(options.color_last_edge);
-					obj->setLineWidth(options.width_last_edge);
-				}
-				else
-				{
-					// Normal format:
-					obj->setColor_u8(options.color_normal_edge);
-					obj->setLineWidth(options.width_normal_edge);
-				}
-				if (is_best_path)
-				{
-					obj->setColor_u8( options.color_optimal_edge);
-					obj->setLineWidth(options.width_optimal_edge);
-				}
-
-				scene.insert(obj);
-			}
-		}
-	}
-
-	// The new node:
-	if (options.new_state)
-	{
-		mrpt::opengl::CSetOfObjectsPtr obj = mrpt::opengl::stock_objects::CornerXYZ(xyzcorners_scale*1.2);
-		string m_name =	"X_new";
-		obj->setName(m_name);
-		obj->enableShowName();
-		obj->setPose(*options.new_state);
-		scene.insert(obj);
-	}
-
-	// Obstacles:
-	if (options.draw_obstacles)
-	{
-		mrpt::opengl::CPointCloudPtr obj = mrpt::opengl::CPointCloud::Create();
-
-		obj->loadFromPointsMap(&pi.obstacles_points);
-		obj->setPose(mrpt::poses::CPose2D(0.0, 0.0, 0.0)); // Points are relative to the origin
-
-		obj->setPointSize(options.point_size_obstacles);
-		obj->setColor_u8( options.color_obstacles );
-		scene.insert(obj);
-	}
-
-	// The current set of local obstacles:
-	// Draw this AFTER the global map so it's visible:
-	if (options.draw_obstacles && options.local_obs_from_nearest_pose && options.x_nearest_pose)
-	{
-		mrpt::opengl::CPointCloudPtr obj = mrpt::opengl::CPointCloud::Create();
-
-		obj->loadFromPointsMap(options.local_obs_from_nearest_pose);
-		obj->setPose(*options.x_nearest_pose); // Points are relative to this pose: let OpenGL to deal with the coords. composition
-		obj->setPointSize(options.point_size_local_obstacles);
-		obj->setColor_u8( options.color_local_obstacles );
-		scene.insert(obj);
-	}
-
-	// Start:
-	{
-		mrpt::opengl::CSetOfObjectsPtr obj = mrpt::opengl::stock_objects::CornerXYZ(xyzcorners_scale*1.5);
-		obj->setName("START");
-		obj->enableShowName();
-		obj->setColor_u8( options.color_start );
-		obj->setPose(pi.start_pose);
-		scene.insert(obj);
-	}
-
-	// Target:
-	{
-		mrpt::opengl::CSetOfObjectsPtr obj = mrpt::opengl::stock_objects::CornerXYZ(xyzcorners_scale*1.5);
-		string m_name =	"GOAL";
-		obj->setName(m_name);
-		obj->enableShowName();
-		obj->setColor_u8( options.color_goal );
-		obj->setPose(pi.goal_pose);
-		scene.insert(obj);
-	}
-
-	// Log msg:
-	if (!options.log_msg.empty())
-	{
-		mrpt::opengl::CText3DPtr gl_txt=mrpt::opengl::CText3D::Create(options.log_msg,"sans",options.log_msg_scale );
-		gl_txt->setLocation(options.log_msg_position);
-		scene.insert(gl_txt);
-	}
-
-} // end renderMoveTree()
-
